Add Car and Person classes, and Main file
A new Car class is created with attributes such as brand, model, color, year of manufacture and actions like start engine and accelerate. A Person class with the attributes such as first name, last name, age and gender is also added. A Main file is also created to initiate instances of these classes and demonstrate the functionality.