Update Library class with book management methods
The Library class has been updated with methods to add and delete books, providing the ability to manage and manipulate the books in the library. The scope of 'books' variable in Library was also modified to public for these changes. These updates also led to minor modifications in the Main program reflecting the new Library methods usage.